requirements.txt
aiohttp
requests
beautifulsoup4
pandas
python-dotenv
psycopg2-binary

fastapi
uvicorn
sqlalchemy
pydantic